refactor /manage directory structure
