+++ src/features/manage/common/utils/convertDifficultyToNumber.ts
