+++ src/features/manage/common/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./convertDifficultyToNumber";
 export * from "./convertJPToENWeekday";
 export * from "./convertEnToJPWeekday";
+export * from "./toSelectColorLabel";
