+++ src/features/manage/components/ManageDayOfTheWeekSwitchButton.tsx
@@ -1,6 +1,6 @@
 import type { FC } from "react";
 import type { Day } from "../../../api/quest/types";
-import { convertEnToJPWeekday } from "../common/funcs";
+import { convertEnToJPWeekday } from "../common/utils";
 
 type Props = {
   view: Day;
