+++ src/features/manage/components/ManageDifficultyCheckBox.tsx
@@ -1,7 +1,7 @@
 import type { ChangeEvent, FC } from "react";
 import type { Difficulty } from "../../../api/quest/types";
 import { Star } from "../common/components/Star";
-import { convertDifficultyToNumber } from "../common/funcs";
+import { convertDifficultyToNumber } from "../common/utils";
 
 type Props = {
   handleDifficulties: (difficulty: Difficulty) => void;
