+++ src/features/manage/components/ManageTimetableCard.tsx
@@ -3,7 +3,7 @@ import { type FC, useState } from "react";
 import type { Day, Difficulty } from "../../../api/quest/types";
 import { ClockIcon } from "../../common/components/icons/ClockIcon";
 import { calcExp } from "../../common/funcs/calcExp";
-import { convertEnToJPWeekday } from "../common/funcs";
+import { convertEnToJPWeekday } from "../common/utils";
 import { ManageProgressBar } from "./ManageProgressBar";
 
 const convertENToJPWeekdayString = (weekDays: Day[]) => {
